refactor bottom navbar and splash screen into components folder
