+++ lib/components/splash_screen.dart
